jjjj

+++ src/parser/execute_command_line.c
@@ -53,6 +53,8 @@ void handle_all_redir(t_shell_control_block *shell)
 
 void	process_command(t_shell_control_block *shell)
 {
+  int built_in_result;
+  
   shell->in_file_name = NULL;
   shell->file_name = NULL;
   handle_all_redir(shell);
@@ -68,14 +70,11 @@ void	process_command(t_shell_control_block *shell)
     dup2(shell->fd_in, 0);
     close(shell->fd_in);
   }
-  if (shell->line_pointer && shell->line_pointer->type == PIPE)
-  {
-    close(shell->arr[0]);
-    dup2(shell->arr[1], 1);
-    close(shell->arr[1]);
-  }
-  if(!execute_built_in(shell, child))
+  built_in_result = execute_built_in(shell, child);
+  if(!built_in_result)
     execute_command(shell);
+  else
+    exit(shell->exit_status);
 	if(shell->in_file_name)
   		unlink(shell->in_file_name);
 }
@@ -92,6 +91,13 @@ void execute_command_line_helper(t_shell_control_block *shell)
     {
       dup2(shell->previous_read_end, 0);
       close(shell->previous_read_end);
+    }
+    if (shell->line_pointer && shell->line_pointer->type == PIPE)
+    {
+      close(shell->previous_read_end);
+      close(shell->arr[0]);
+      dup2(shell->arr[1], 1);
+      close(shell->arr[1]);
     }
     process_command(shell);
     exit(0);
